script/requirements.txt
pyprelude
pyyaml